# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEP C2M Service API",
        "description": "VEP C2M Service API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "",
            "description": "clouddev"
        },
        {
            "url": "",
            "description": "sit"
        },
        {
            "url": "",
            "description": "uat"
        },
        {
            "url": "",
            "description": "prd"
        }
    ],
    "components": {
        "schemas": {
...
            "ResetPendingMeetingNotificationStatus": {
                "type": "object",
                "properties": {}
            },
            "NotiSqsMessageBodyDto": {
                "type": "object",
                "properties": {}
            },
...
            "SetHiddenRecordDto": {
                "type": "object",
                "properties": {}
            }
        }
    },
    "paths": {
...
        "/notification/getNotificationRecordByMeetingId/{meetingId}": {
            "get": {
                "operationId": "NotificationController_getNotificationRecordByMeetingId",
                "parameters": [
                    {
                        "name": "meetingId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/notification/notification/fairs/{fairCode}/fiscalYear/{fiscalYear}/recommendations/buyers": {
            "post": {
                "operationId": "NotificationController_findRecommendedExhibitor",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fiscalYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/resetPendingMeetingNotificationStatus": {
            "post": {
                "operationId": "NotificationController_resetPendingMeetingNotificationStatus",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPendingMeetingNotificationStatus"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/mock-send-sns-msg": {
            "post": {
                "operationId": "NotificationController_postSnsMessageBody",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotiSqsMessageBodyDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/c2m/fairs/{fairCode}/hidden-records": {
            "get": {
                "operationId": "HiddenRecordController_getHiddenRecords",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "HiddenRecordController_setHiddenRecord",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetHiddenRecordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/hidden-records/{userId}/not-interested": {
            "get": {
                "operationId": "HiddenRecordController_getNotInterestedByFairList",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}